+++ CHANGELOG.md
@@ -4,7 +4,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-Feature: Add `bulletProps` and change type of `statusMessage` in `StatusWithBullet`.
+## [2.31.0] - 2019-10-01
+
+Feature:
+- Add `bulletProps` and change type of `statusMessage` in `StatusWithBullet`.
 
 ## [2.30.0] - 2019-10-01